Adds support for rulesets
This adds support for `/global-event-rules/rulesets` endpoints, which are part of the "Rulesets" feature: https://support.pagerduty.com/docs/rulesets#section-create-multiple-rulesets-and-assign-teams

Rulesets replace Global Event Rules so it seems these are the endpoints we should use going forward.

Note that these endpoints do not have documentation available at this time and are subject to change.